Change of organization of dataset folders
Previously, it was a mix of genuine and forged signatures within the train, as well as the test folder.

Now, the signatures are further sorted into "Real" and "Genuine" folders, for it to work better with the iterative access traversal code as seen in the "Saving The Features" portion of the skeleton code.